pyproject.toml
[tool.poetry]
name = "isidore-referentiels"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.1"
requests = "^2.32.3"
rdflib = "^7.0.0"
urllib3 = "^2.2.2"

[tool.poetry.scripts]
isidore_referentiels = "isidore_referentiels.main:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
